requirements.txt
Flask==3.1.0
websocket-client==1.8.0

